+++ CHANGES.rst
@@ -4,7 +4,11 @@ Changelog of lizard-progress
 4.1.11 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Only perform a max distance check when the Location is not newly created.
+
+- Use midpoint for creating metparser Locations.
+
+- Use midpoint for TOO_FAR_FROM_LOCATION change request.
 
 
 4.1.10 (2018-06-20)