+++ CHANGES.rst
@@ -6,6 +6,12 @@ Changelog of lizard-progress
 
 - Hotfix progresspar project dahboard.
 
+- Only perform a max distance check when the Location is not newly created.
+
+- Use midpoint for creating metparser Locations.
+
+- Use midpoint for TOO_FAR_FROM_LOCATION change request.
+
 
 4.1.10 (2018-06-20)
 -------------------